Merged PR 538749: Add a preliminary Rush frontend to BuildXL
Adds a preliminary Rush front-end to BuildXL. All fundamental pieces in place. This frontend should be able to build sp-client repo by just placing a bxl config file at the root.
From  an architecture point of view this is very similar to the MSBuild one: an out-of-proc tool builds a project graph that is then serialized into JSON and used by the frontend to schedule pips. Most logic is done by the existing rush-lib utilities.
Pending:
- Some basic tests coming up
- There are still many details that are sp-client specific and need to be generalized. Corresponding TODOs should be in place.