Merged PR 540786: QTest: Generate InputsManifest file to optimize Sandbox file copying
###Description
QTest sandboxing involves multiple copy operations to make sure the test sandbox is not dirtied after each test run.
Before this change, QTest was performing some extra file copy operations which were not efficient.

1. To pass QTestInputs from Office builds to QTest Sandbox, the SDK would copy the inputs into a single sealed directory before passing to QTest executable.
2. Inside QTest executable, QTest would make a clean copy of the required test files to then copy into the test sandbox for each run.

This change avoids both 1 and 2 above. This is achieved by creating a single InputsManifest file in QTest SDK to pass to the QTest Executable. This inputs Manifest file is then used as the clean state of test input files to be copied to the test sandbox for each individual test run.

This change reduces QTest Sandbox file copy operations by at least 1 set of test files.
For Office builds that pass QTestInputs, this change includes an additional saving of one full set of copies.

###Updated QTest Nuget
Includes fix for rerun script.

###NOTE
Office uses QTestDirToDeploy argument for some of their tests. When QTest SDK attempts to enumerate files in this directory, it returns 0 files, failing to populate the inputs manifest.
For this reason, we are using --copyToSandbox as a fallback mechanism for these cases.

###Testing builds
[Office WAC](https://cloudbuild.microsoft.com/build/78b0f4d7-06e6-492c-8c32-45833917c607?bq=Bxl_OfficeTenant_Wac_Release)
[BXL Buddy](https://cloudbuild.microsoft.com/build/2fc3fe0f-342a-49de-903f-51ef84e73fca?bq=BuildXL_Internal_Buddy)

Related work items: #1678727