Merged PR 540381: Limit place file operations under put throttle
Currently put file operations are under a throttle with a semaphore implemented. We want to integrate place file operations into this put file throttle, and am accomplishing this sharing the semaphore for put files with place file operations.

This results in a maximum number of concurrent put and place file operations, which is set by a configuration value "MaximumConcurrentPutAndPlaceFileOperations", defaulted to 512.