Merged PR 541702: Fix missing version update of .net core 3.0 to 3.1
We went back between 3.0 and 3.1 a bunch of times in the past due to windows defender accidentally marking the 3.1 bits as containing suspected code. It seems in the back and forth these small changes were accidentally left behind.

This was causing errors when trying to run the xunit unittest using resharper in vs 2019 with our generated solution and .csproj files.