Merged PR 541647: Only log ProcessFingerprintComputation if it has a value
Only log ProcessFingerprintComputation if it has a value.

From Kusto query I found that the cache miss analysis results of the similar pips mentioned in https://dev.azure.com/mseng/1ES/_workitems/edit/1682109 were actually comparing the data from failed pips.

The previous pip failed before the observed input is observed, so the path of this input was not added to the pathset. In the later build which the pip was success, the observed input of the pip was added to the pathset. That's why we see the new added path in the analysis result.

With this change, cachemiss analysis won't perform for fail pips. We don't know at which point the pip will fail. The ProcessFingerprintComputationData can be incomplete. Performing cache miss analysis with incomplete data can give misleading result to users.

Related work items: #1682109