Merged PR 541816: Improve XUnit integration on generated solution for Resharper
We were missing some things for resharper integration to work.
* The testhost is moved to the base xunit from qtest only since VS needs the same reference.
* Generated UnitTest projects now generate pacakge references to the test host.
* Fixed unittest detecton logic that was only looking for dotnet, not dotnet.exe

While I've worked through a few obstacles, it still doesn't fully work but we are closer. My current hypothesis is that because we don't run some of the build steps and the extension uses `dotnet vstest` do run some stuff we need to beef up faking some msbuild targets.