Merged PR 540931: Fail gracefully if PushFile fails because server is unavailable
Unlike `CopyFileAsync`, `PushFileAsync` implementation never respected headers to gracefully fail if the server is unavailable.

This explains why we're seeing so much more failures for push files during service version rollouts or config changes compared to the failures because of copying the files from the machines.

This change introduces `PushFileResultStatus` that can be one of 3 values: `Success`, `Unavailable` and `Rejected`.
And now all 3 cases are considered to be "successful" cases, just because we don't treat this operation as a high critical one.

The tracing was changed to see why the operation has failed or even when the push was fully successful only in one case (inner ring) but not in another.

Related work items: #1693860