Merged PR 542089: Add tracing to redis operations
In RedisGlobalStore we fire off redis operations that reach RedisDatabaseAdapter and RedisBatch, which then trace in a FireAndForget task if the operation fails. However, these traces do not explicitly state whether it was the primary or secondary redis database.

To fix this, we now will have a database name for each RedisDatabaseAdapter, which it passes to the RedisBatch to trace before a FireAndForget task.
For example: "PrimaryRedisDatabase Unhandled exception in fire and forget task for operation ..."

The FireAndForget task is changed to possibly take a tracePrefix, when we want to add at the start of the trace for extra information.

Each RedisDatabaseAdapter now also takes in a configuration, with new configuration values setting whether the adapter traces normal redis operation failures and transient failures.