Merged PR 542638: Do not scrub deployment dir for Rush related tests
Our rolling has been failing non-deterministically with:
[error]Catastrophic BuildXL Failure.
[error]Build:[0.1.0-20200323.5][refs/heads/master:6b7ba0fa8b03935d5422c1ca6a055945dd85a31f].
[error]Exception:BuildXL.Utilities.BuildXLException: Delete directory contents failed after exhausting retries because some files or directories still existed in the directory. This could be because an external process has a handle open to a file or directory. Directory: D:\bld\bld\1\s\Out\Objects\0\8\3zd42dknx2eh6y25bqqvv1q8\testRun\rush

I haven't got to the root cause, but it seems like executing rush (as part of the test) does not play along with scrubbing the test deployment for stale files. One reason could be that temp files get placed under the npm-installed rush deployment as part of the tool execution.
This PR adds an (unsafe) option for tests to avoid creating a fully sealed with scrubbing on for their deployment.