Merged PR 542602: Pin Threshold Explicit
This PR is the second rollout for making pin threshold explicit. Previously, we were compute pin thresholds based on a set of different pin configurations (pin risk, file risk, etc), but now we will have a new config value that allows us to just set the number for pin threshold without computing it.

As well as removing most pin configuration values, removed all of pinCaching, and ContentAvailabilityGuarantee. The respective pin tests will take the new config value for PinMinUnverifiedCount, and PinConfiguration no longer is ever null, so we do not check for that case.