Merged PR 543058: Setup infrastructure to run CloudTests tests that run during our CI as well as can run in our PR validation.
This PR will finish the instrastruture needed to run tests in cloudtest from our repo.
* There is an optional CloudTest validation step in the PR. Currently called `CloudTest validation (Gvfs)`.
  - If we do anything to the change detection logic we should manually run this.
* If you change anything under the CloudTest folder a sibbling validation will be automatically kicked off and is marked required called `CloudTest validation (Gvfs) - Required`
* There is a validation step in our CI between the Dogfood and Canary stages called `Cg: CloudTest validation (gvfs)`
  - This one is currently marked as optional until we have enough confidence in the signal and can make it required.

The infrastructure has been set up in that it is not just for Gvfs validation. Future CloudTest tests can simply be added as a sibbling under the `private\CloudTest` folder following the gvfs tests pattern and following the user guide on `http://aka.ms/CloudTest` on how to edit the xml files.
If we add more  I would recommend naming the queue and the jobs in the pr validation and CI to reflect this.