Merged PR 543725: Fix crash in PerformanceCollector on macOS
The crash was happening because GetDiskCountersMacOS() builds up its results in a list. When there is an exception getting drive information that list is not added to, and thus there's a mismatch in the count of drives between what was initially discovered and recorded in m_drives vs. the collected sample.

The change keeps the drives consistent, adding an IsValid field to the struct to track when the value should be ignored on error.

Also a slight refactoring for the UT hook.

Related work items: #1700685