Merged PR 543781: separating blobs into two redis instances
We are attempting to decrease the load on one of the redis instances (primary), by sharding the blobs, and placing half of the blobs on the secondary redis instance.

We decide which redis instance to use based on the first byte of the content hash, whether the first byte is greater than or equal to 128.

There is also the chance, secondary redis instance is null, in that case, we send all blobs to the primary.