Merged PR 543674: BlobCentralStorage TryGetFileCoreAsync adding retries
Previously we would iterate through all blobCentralStorage container and shards, and determine whether getting the blob from that container and shard was successful. If the result could be retried, we didn't retry, we just moved onto the next iteration of container and shard. If the result could not be retried, we break out of the for loop.

In this PR, we add a blobStorageRetryStrategy that retries between the default exponential intervals. The retry strategy executes getting the blob from each iteration of container and shard, and if an exception is thrown, we determine whether it is transient and possible for retry.