Merged PR 544767: Add BuildXL workingset usage to TotalAvailableRam at Execution phase
We were using the available memory at the beginning of execute phase as the total ram. This total ram is then used to calculate for forecasting. When looking at the telemetry data, I noticed that graph construction can take up-to 50gb. This memory is not released during the execution phase unless there is memory pressure. However, we might limit the scheduler by using higher memory semaphore due to the low totalrammb.