Merged PR 544943: Disable SOD when running un-sandboxed
Disable SOD when running un-sandboxed to avoid sideband file read / write exceptions. The sideband files are not filled when running un-sandboxed to begin with, this disables SOD in that case - before BuildXL would fail with an error indicating the missing SOD file for a PIP that uses shared opaque directories explicitly in its DScript definition.