Merged PR 545063: Fix crash when error is logged due to failure to filter pip output
Error handling paths in TryFilterAsync() can log an error if there's an IOException. We see this happen in practice when machines run out of disk space. Generally this doesn't cause an issue because the pip would naturally fail in that circumstance, but if a pip gets retried an error is logged even though the pip may go on to fail, causing the crash. I found another instance where an error can happen without the pip failing in the logic that replays warnings from the cache.

Related work items: #1703051