Merged PR 545335: Use INVALID_HANDLE_VALUE explicitly when calling GetFileInfo during failed NtCreateFile
The handle value can be considered "invalid" when NT_SUCCESS is false (there could be plenty of reasons). Unfortunately, we assume that the value will be INVALID_HANDLE_VALUE, and use that handle value to determine if a path is file or directory.

This change simply use INVALID_HANDLE_VALUE explicitly.

This change is needed to unblock Azure compute when running with tracker in QuickBuild.

Related work items: #1705380