Merged PR 545720: Expose augmented file accesses in IDetoursEventListener
This is required for some QuickBuild features to continue to work when using shared compilation, as these need to differentiate between true MsBuild accesses and augmented accesses representing what the compiler would have made.

Made this a breaking change per @<Serge Mera>'s suggestion as QuickBuild is the only consumer.