Merged PR 545702: Do not publish empty fingerprint store
Dominorunner invokes a nuget only session before the actual build sessions. This nuget only session creates an empty fingerprint store which get published at the end of the session. Since this is a session always happen during the build right before the actual build sessions, So the empty fingerprint store it publishes is almost always the latest cache entry for the following actual build and be fetched. The empty fingerprint store can't be opened and no cache miss analysis can be performed.

Related work items: #1681704