Merged PR 545154: Change distributed delete to trace information on all deletes
In the previous distributed delete we had no tracing when the delete verb was called. Also, because we only had DeleteResult implemented, we couldn't retrieve information for all the delete results. To hack around this, we took the highest priority delete result code.

In this PR, we created a DistributedDeleteResult class that derives from DeleteResult, and includes a mapping from MachineLocation path to its corresponding delete result. Then when we trace, we output all the results from different machine locations, and also trace any errors or other information when needed.

For size, we used to accumulate the evicted size. But now we will display the content's size, and the total evicted size can be determined by how many successful deletions happened and the content size.

Pinned size is not needed for delete result so it is removed.
RedisBatch added an extra trace in its fire and forget call to trace primary/secondary redis database.