Merged PR 546549: Update StackExchange.Redis to 2.1.30
This PR updates `Redis.StackExchange` library to version 2.1.30 to include a fix that supposedly solves the problem with the inability of a process to connect to Redis instance.

This is the PR that this version includes: https://github.com/StackExchange/StackExchange.Redis/pull/1374

We do have many cases when the process fails to connect to on Redis instance and after the service restart it immediately (within seconds) successfully connects to it.

Related work items: #1697377