Merged PR 547049: Do not retry some exceptions while uploading symbols
When we upload symbols, we sort of expect to see DebugEntryExistsException. There is no point in retrying a call, because one can only add items to the symbols publish request, ie., once a file is there, it stays there. Also if there are enough collisions in a build, our retries trigger DDOS protection on the Artifact Services endpoint -> failed builds.

Related work items: #1660276, #1706582