Merged PR 547262: Enable C# compiler's strict mode
The C# compiler supports a strict mode, the mode that was design to emit extra diagnostics for existing code and not break existing customers. The classical example is definitely assignment checks when all the public members are removed because the type is used via reference assemblies (we hit this case actually with `RelativePath`), or when a value type can be used for doing locks (we hit it as well).

Related work items: #1710763