Merged PR 546297: Tracing for unobserved exceptions
Currently we are tracing a bunch of well known exceptions to Kusto such as RpcException. We know these exceptions are being thrown and currently do not want the process to be exited, or the exceptions to keep tracing. This PR is just to mitigate the tracing until we eventually invest in figuring out why the unobserved exceptions' root causes.