Merged PR 548502: Add packages.config file to allow test discovery by VS
To run the tests in VS some requirements must be met:
1. The solution (at least for me) must be generated with `/vs` and with `/q:DebugNet472`
2. The test project should have `packages.config` file that references `xunit.runner.visualstudio`.

This PR adds package configs for all the projects that were missing them (the cache tests have them for some time already).

Please note, that once the PR is merged and you generated the solution already with `/vs /q:DebugNet472` you don't even need to regenerate the solutions in order to run the tests.

Related work items: #1711514