Merged PR 549295: Allow attribute writes on hardlinks
A while ago, a Windows update made it so you couldn't hardlink if allow attribute wristes was set to false. This led us to default to set this value to true. However, for caches that already have the content, there was no fix to address this other than deleting the cache.

This PR makes it so that if we fail a hardlink, we verify that this is set to true or set it otherwise, and retry the hardlink.

Tested this locally.

Before my change:
1. Put a file to the cache.
2. Manually set allow attribute writes to false
3. Try to place via hardlink. Fails and falls back to copies.

After my change:
1. Put a file to the cache.
2. Manually set allow attribute writes to false
3. Try to place via hardlink. Succeeds via hardlink.