Merged PR 550771: Enable directory symlink processing under a flag
Enable directory symlink processing for the Windows case under a flag. This change may become eventually moot and subject to removal after the ongoing detours changes start processing directory symlinks appropriately.
This PR is essentially trying to unblock the scenario for Rush-based repos, which are not creating directory symlinks during the build, but most read accesses go through dir symlinks.
The approach here follows what Mac is doing: any access that go through directory symlinks is resolved so the real path is used downstream, and additionally, a read access on every dir symlink is produced so underbuilds are avoided.

(note to @<Kristijan Simic> , observe that one of the nice parts of doing this on the manage side is that we can keep a cache across processes, which saves tons of IO since the same paths tend to show up many times in a given build. You may want to consider keeping some of that even after the detours work is done? Not sure if possible)

Related work items: #1697595