Merged PR 552319: Don't link graph files to logs directory on workers in distributed builds
In distributed builds the graph files in the logs directory are redundant for the workers since they're the same as what's on the master. The XLG on the master has all of the forwarded worker events so consuming the worker's XLG isn't really a use case anyway.

The value add of this change is to reduce the size of the log directory on distributed build workers