Merged PR 552398: Allow configuring to preserve the pathset casing at the pip process level.
On Windows, the path set gets serialized in all uppercase as an attempt to avoid having different path sets for the same weak fingerprint, in detriment of cache efficiency.
This causes particular trouble in the JavaScript world, where many tools expect casing to be preserved.
This PR adds a configurable option at the pip process level to preserve path set casing and follows PathTable casing. It also makes the rush pip constructor turn on that option for all rush-scheduled pips.

Related work items: #1697633