Merged PR 551568: Enable non-nullable types in FileSystemContentStoreInternal
TLDR; The PR removes a bunch of old code and moves old tracing like `PutFileCall<ContentStoreSessionTracer>` into the extension methods.

This PR:
* Enables non-nullable types in `FileSystemContentStoreInternal`
* Removes some old code that was left after removing old redis implementation
* Removes CAS versioning because the last time the version has changed was more than 2-3 years ago
* Moves all the tracing to extension methods, i.e. no `PutFileCall<Tracer>` in `FileSystemContentStoreInternal`, I haven't moved to the new logging, but I extracted the logic into the extension methods to make sure the behavior has not changed. Now the tracing code is not in our face (which is good) and it can be easily changed without changing the logic.

I was thinking to introduce a new base class `StartupShutdownNullFriendly` or to change an existing one that will simplify null-friendliness, but I decided to do that separately to avoid making too many behavioral changes.