Merged PR 553126: Add cap for number of errors/warnings sent to Azure DevOps listener
Sets a limit of 500 warnings and 500 errors to be emitted as issues by the DevOps listener. This is to prevent slow rendering of the build results page.

A message of the same type is added to let end users know the number of errors/warnings rendered has been truncated.

Related work items: #1717244