Merged PR 554167: [Rush] Add option to use shrikwrap-deps file instead of tracking the actual project dependencies
When up-to-date, shrikwrap-deps.json is a legit witness of a project transitive dependencies. Add an option to use this file instead of tracking all dependencies. This option is safe when running in a lab environment, where rush install/update is always run before bxl.
This brings a couple of benefits:
* A lot less tracking/hashing, so perf may be better
* Make cache hit/miss work with compatible dependency changes. In this scenario, Rush allows projects to consume dependencies which don't exactly match the rush lock file, but are considered semantically equivalent.

Related work items: #1697633