Merged PR 553919: Add copyright text to .editorconfig and update the codebase with new rule
Starting with Visual Studio 16.6 the .editorconfig can contain a copyright text that is enforced by the IDE.

This PR adds the new option and also I generated the full solution and fixed all the known violations of our copyright header.

The feature is available only in 16.6 but you can stay on the older version without any issues, just this feature won't be available to you.