Merged PR 556223: Implement asynchronous copy on pin
When we originally discussed this work item we were thinking of having another threshold. For instance, when unverified pin threshold is 8 and the number of replicas is less than, let's say, 6, we'll copy asynchronously.

But we know that the eviction storm can cause issues even when we have the max number of locations (let's say 8), so instead of having another number, I decided to add just a flag: if the flag is set then we trigger an asynchronous copy every time we're satisfying the pin based on the number of locations.

Related work items: #1727961