Merged PR 556495:  [linux-sandbox] Handle the case when a surviving child process crashes
Tracking of child processes on Linux is implemented by providing an `on_exit` callback; the problem is that `on_exit` callbacks don't get called when the process crashes.  Consequently, if a child process crashes, the Linux sandbox (concretely, `SandboxConnectionLinuxDetours`) never receives that notification and continues to wait forever.

The solution implemented in this PR is to start polling all surviving child processes after the root process exits.

Observe also that we already have a reliable mechanism for detecting when the root process exits (even if it crashes): the root process is spawned using the C# Process API which guarantees that the `Exited` event is fired when the process exits.