Merged PR 556372: Consolidate copy ordering in MachineList
Consolidate copy ordering in MachineList. This is an effort to simplify our code and have everything in one single place.

This means that the distributed copier is no longer responsible for ordering locations in any way, but instead should attempt copies in the order they are given to it.

Related work items: #1723515