Merged PR 558827: Push file clean-up and perf optimization
This PR has 2 main changes:
1. Remove non-push version of proactive copy (the push version is used ubiquitously in prod and in cbtest)
2. Change PushFile api to take a stream instead of taking a callback that returns a stream.

The second change is very important because it fixes the following issue during proactive copy: proactive copy is a long-running operation and the time between obtaining the hash and the time when the "hash" is pushed can be quite substantial (like hours). And in some cases, the push can fail because the hash is already evicted.
The issue with the old approach is that the client sends the request to the receiver first and only after getting acknowledgment from the receiver, it tries to open the stream that can fail (it actually fails like ~2M times in CB Test a day!).

It was possible to fix the issue just by moving the callback invocation to obtain the stream before making the first call to the server, but I think the callback makes the API more complicated than needed.