Merged PR 559391: Ensure VM semaphore gets disposed and retry semaphore creation
We are trying to address the issue:
```
2020-06-14 01:30:41,531 ERROR Process stderr:
Semaphore 'd:_dbs_el_ods_Build_DominoTemp_Detours_3E47D57795F9E19F-07dd9851-cbfc-42df-bbce-c5caf3d3bb61.tmp' for counting Detours messages is already opened

```
From the code path and the fact that VM doesn't retry, the error seems to be impossible, given that GUID is part of the semaphore's name.

The changes try to address this issue by ensuring that semaphore gets disposed when executing in VM, and retry the semaphore creation (with different names).