Merged PR 559612: Move shared opaque output classification to SandboxedProcessPipExecutor
Consider the following scenario:
* Pip A creates a file with path p
* Pip A deletes a file with path p
* Pip A creates a directory with path p
* Pip B probes p

p is first considered a shared opaque candidate, so it is added to a shared output structure that then is used to do file monitoring analysis. But in the postprocessing phase under PipExecutor the path is actually discarded since the final artifact is a directory.

This introduces a bug, where Pip B gets a filemon violation if Pip A executes. This violation does not happen if Pip A gets replayed from the cache.

This PR fixes the issue by moving shared opaque output classifications (file present/absent or directory) which was happening under PipExecutor to the SandboxedPipExecutor. In this way we can discard paths whose final artifact is a directory earlier and not make it part of the shared output structure.