Merged PR 558474: Handle RedisConnectionException when failed to connect to nonexistent instance
When a redis instance does not exist, and we attempt to connect to it we get a RedisConnectionException. The unique part of this specific message compared to other RedisConnectipnExceptions that may have been caused to config issues is that the message has "UnableToConnect" contained in the message.

We now look for this message in the RedisConnectionExceptions, and if this is the case, we do not increment the connection error counter, which determines whether we need to restart the connection multiplexer. We also mark this error as non transient so do not perform further retries, and trace the error message.