Merged PR 559761: Add an option to disable composite SOD filters
To do A/B testing, we need a way to control whether the filters are applied or not. Ideally, this would be an unsafe option, but making it unsafe would mess up with caching (which in turn would affect the results). Once we are done with the A/B testing, we'll rollback this change.