Merged PR 559827: Rename whitelist to allowlist
This maintains compatibility with configurations and build graphs using the old name but generally renames the feature.

There are a few straggling instances in the macOS sandbox code to update but this was getting pretty large so I wanted to send it out quickly to avoid risk of creating merge conflicts with renames.

Related work items: #1737635