Merged PR 559626: Prevent race condition for hibernating session and Quotakeeper
Previously on shutdown of the server we would hibernate sessions and shutdown the sessions. During shutting down of the sessions the content pinned would be freed, and allowing a small window for the quota keeper to evict content that was meant to be pinned before the quota keeper gets called to shut down.

To prevent this from happening we will call the quotakeeper to shut down before we hibernate sessions. To prevent concurrent and duplicate shutdowns of the quotakeeper we wrap the code responsible for shutting down the quotakeeper with a semaphore, and if the shutdownCompleted flag is true, we do not perform another shutdown.