Merged PR 561630: Update AnnotatedSessions to count unclean exits for otherwise unexplained failures as internal errors
This came up because Cosine was seeing some failures where bxl.exe was mysteriously exiting. It turned out to be fairly common so I think we should track it in our internal error triage until we get to the bottom of it.

Office has proportionally fewer, but still quite a few. All of the office instances of this are marked as DeadClientOrLostConnection. I'm not sure what the root cause for them is yet.