Merged PR 561070: Fix halt for singlemachine builds for the decouple materialize feature
We have been using this feature for Cosine and some Office builds in CloudBuild with no issues. The bug is related to the single machine builds where we pass /enableLazyOutputs-. If you pass it on single machine builds, MaterializeInput step is skipped. We were incrementing the MaterializeInput slots; but we do not decrement because the step is being skipped. Hence, the hang was occurring on the ChooseWorkerCpu dispatcher.