Merged PR 561361: Embed debug information (pdb) into assemblies
This change adds another "DebugType" enum to C# compiler that allows to embed debug information into the assembly itself and also changes the default for bxl repo (maybe these steps should be separate though).

This is quite an important change because it means that regardless of the consumers of our bits, the line information is always preserved and the stack traces will always have the right line numbers.

This feature simplifies the consumption of bxl bits, because by default when the packages are referenced via package references feature, msbuild will only include the dll's and not pdbs.

For instance, CloudBuild now has a special code in there codebase to explicitly include bxl pdbs to deploy them alongside with the dlls in CaSaaS deployment folder. But unfortunately, casaas is not the only consumer of bxl packages. For instance, QuickBuild also consumes them and lack of pdbs there means that the line numbers are missing when the error is happening in QuickBuild builds.

Embedding the pdbs actually saves overall space and the new nuget packages are actually smaller than they were before.

Stuff to check before checking this in:
1. Check that casaas crashes still have the right line info.
2. Check on Linux
3. Validate in office queues

Related work items: #1741496