Merged PR 564119: PipCompletionData is retrieved by both pipId and executionStep
if a worker takes long to execute a pip build request, the orchestrator can give up and execute that request on another worker. Then, materializeOutput request will be sent to all workers during metabuild including the initial machine. Because pipCompletionData is only indexed by the pipId, we do not want to reuse the same pipCompletionData for the second materializeOutput request. With those changes, pipCompletionData is specific to both pipId and executionStep.

Related work items: #1748632