Merged PR 564350: Optimizing MachineIdSet
For a long time, we had seen that `MachineIdSet` implementation was not optimal and was causing a lot of memory traffic.

The class is immutable but it was creating `HashSet` on every `SetExistence` call.

This PR changes the implementation and move is to use `ImmutableArray<ushort>` instead, with the following benefits.

1. Reduced memory traffic on master machines
The memory traffic dropped roughly about 7-8Gb per 1 minute of normal operations in WDG stamp.

2. It is faster as well (even though the perf benefits are not really visible in the real scenarios, or it's hard to measure them).

I'll post the benchmark results in the comment, otherwise, the description is too long.

The most benefits we're getting when we're changing only one location. But this is actually a common case. I also checked the case when a bunch of locations were changed at once to see that there was no regression in perf because of the quadratic complexity of the `SetExistence` implementation.