Merged PR 562297: Fix DW DFAs for temporary files under shared opaques
Pips can create temp files under produced shared opaque directories. There is an issue when two pips decided to create the same temp file (ie., collide on the file naming). Currently this leads to DW DFA.  It's not feasible for a customer to change how those pips work / select temp file name, so we are relaxing the treatment of such files for cases where there is a dependency between the pips.

Related work items: #1738652