Merged PR 565707: [symbols] Add support for output directories
Our symbol indexing logic is creating the indexing process pip that takes care of extracting the "keys" required for the actual symbol publishing. Since it's a process pip, the metadata file is replayed from cache so we don't need to materialize/re-index the inputs. This does not work for SODs because SOD's content is not a part of weak FP. There are several ways we can add support for indexing SODs:
1. Add a new renderer for PipData (similar to Artifact.vsoHash(xxx)) that would translate directory artifact into the list of files.
2. Add an new pip in front of the indexer pip that would dump the content of the SODs into some file that the indexer can take as an input.
3. Let indexer pip talk to api server so it can get SODs content AND enumerate the SOD roots to force proper caching.
4. Make the calls to the api server a part of strong FP.

Due to various reasons, this PR implements the second approach.

Related work items: #1743877