Merged PR 566796: Always update tracker for async copy on pin
This PR has 2 set of changes:

1. It creates a test that covers async copy on pin in way more details. It checks that the lifetime of the pin and async copies are not linked to each other and it also checks that the location is registered once the async copy is done.
2. It changes the behavior of `CancellableOperationContext.Dispose` method.
This is actually the root cause of the cancellations we were seeing in async copies on pin.

`CancellableOperationContext` effectively links a given context together with another cancellation token. The main use case for that is to have a special context that will notify about a shutdown cancellation.

But in some cases, the call to `Dispose` method was cancelling a context as well.

Here is the scenario: `ContentSessionBase.PinAsync` was creating a "linked" context and was passing it to `PinCoreAsync`. Because of the behavior I mentioned above, the token from the context was cancelled when the pin is done.
This was causing the cancellation of the async operation before.

Related work items: #1755629