Merged PR 566135: Update distributed pin result telemetry
Adding specific error codes for different scenarios of distributed pin results, and adding replica counts to the traces.

Local content code does not exist, and all pin results that are not traced in the form #replicaCount/Code can be assumed as a local pin lookup.

Copy Async code just implies to the user we attempted to copy the content asynchronously, but isn't the propagated result from the async copy, since that task is done FireAndForget. It just lets the user know that the content had enough replicas but not above the threshold for async copy.