Merged PR 565656: Cancel redis batches when connection multiplexer is closed
This PR adds support for cancelling pending redis batches when the connection multiplexer is closed.

This change should address the issue that can some operations to hang forever. For instance, we had seen that when the multiplexer is closed some batches, like onces responsible for getting checkpoint information may never be finished.