Merged PR 567880: Roxis
This PR merges the currently experimental version of Roxis against master branch. This is required in order to make progress on testing Roxis in a CBTest stamp.

Note: most of this code is non-productive and hence shouldn't be held to the same standards. We are just making fast progress towards getting a POC working.

Things that do need to be reviewed are all changes to already-productive cache components, simply to ensure the new changes aren't breaking.

Related work items: #1746843