Merged PR 568395: Fix retry crash with pip failing without logging an error
Previously, when preparation failed the pip was being marked as a failure prior to creating RetryInfo and going through that process. That caused the pip to be marked as a failure without any errors being logged. This changed reorders the check to give precedence to processing the retry.

Related work items: #1757365