Merged PR 567957: Add safe source rewrite relaxation policy
This PR introduces a new policy that allows rewriting sources/alien files unknown to the build when this doesn't affect correctness guarantees. The policy focuses on dynamic writes (under shared opaques) since pip graph construction restrictions are still in place and will block most of these cases.

The new policy will allow rewriting an existing file when a) all readers are ordered after or before the write, which guarantees they all see consistent content or b) bxl can assert the rewrite is a same-content one, in which case even unordered readers are allowed.

Pips that are allowed to (safely) rewrite an existing file are cached as usual. However, the rewritten file is not flagged as a shared opaque, since the file existed beforehand and therefore we don't want to delete it on the next build (since readers may be present and the deletion will naively happen before the build begins). Additionally, the rewritten output will not be hardlinked from the cache since we want to keep ACLs so the rewrite can happen again.

For reviewers: even though there are many files changed, most of them are about passing an extra mark to the cache that indicates the file is a source rewrite. This cross cutting change involves many files, but those changes are straightforward. The relevant changes are in FileMonitorViolationAnalyzer (for determining when a rewrite is safe), SandboxedProcessPipExecutor (for flagging a file as a source rewrite), FileContentManager (for placing the file as copy), Scheduler (for not flagging rewrites as shared opaques), and FileArtifactWithAttributes (for adding the extra flag reusing the same space). I can also schedule a walkthrough if needed.

Related work items: #1742856